requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
pydantic==1.11.1
python-dotenv==0.21.0
